nebius_server/requirements.txt
fastapi
uvicorn[standard]
lerobot[smolvla]
torch
torchvision
pillow
numpy
httpx
python-multipart
